+++ Scripts/GameData/IncrementalFloat.cs
@@ -4,7 +4,7 @@ public struct IncrementalFloat
     public float baseAmount;
     public float amountIncreaseEachLevel;
 
-    public float GetAmount(int level)
+    public float GetAmount(short level)
     {
         return baseAmount + (amountIncreaseEachLevel * (level - 1));
     }
