+++ Scripts/GameData/IncrementalInt.cs
@@ -4,7 +4,7 @@ public struct IncrementalInt
     public int baseAmount;
     public float amountIncreaseEachLevel;
 
-    public int GetAmount(int level)
+    public int GetAmount(short level)
     {
         return baseAmount + (int)(amountIncreaseEachLevel * (level - 1));
     }
