+++ Scripts/GameData/IncrementalMinMaxFloat.cs
@@ -4,7 +4,7 @@ public struct IncrementalMinMaxFloat
     public MinMaxFloat baseAmount;
     public MinMaxFloat amountIncreaseEachLevel;
 
-    public MinMaxFloat GetAmount(int level)
+    public MinMaxFloat GetAmount(short level)
     {
         return baseAmount + (amountIncreaseEachLevel * (level - 1));
     }
